requirements.txt
iniconfig==2.0.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.4
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
PyYAML==6.0.1
